# tt/node/swagger.json
{
  "info": {
    "title": "fusion-swagger",
    "version": "1.0.0",
    "description": "internal-server swagger"
  },
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/users/{login}": {
      "get": {
        "summary": "GET user by login",
        "parameters": [
          {
            "in": "path",
            "name": "login",
            "schema": {
              "type": "string"
            },
            "description": "user's login"
          }
        ],
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "user"
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "summary": "GET filtered and sorted user list",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "filter": {
                    "type": "object",
                    "description": "filter for where",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "notRole": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      }
                    }
                  },
                  "sort": {
                    "type": "array",
                    "collectionFormat": "multi",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "id",
                      "asc"
                    ]
                  },
                  "description": "sorting parameters"
                }
              }
            },
            "in": "body"
          }
        },
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "user"
          }
        }
      }
    },
    "/api/users/editUser": {
      "put": {
        "summary": "edit user record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/UserModel"
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "user updated"
          },
          "500": {
            "description": "error message"
          }
        }
      }
    },
    "/api/users/adminChange/{id}": {
      "put": {
        "summary": "edit user record",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "number"
            },
            "description": "user db Pkey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": null,
            "schema": {
              "$ref": "#/definitions/UserModel"
            }
          }
        },
        "tags": [
          "users"
        ],
        "responses": {
          "200": {
            "description": "user updated"
          },
          "403": {
            "description": "invalid token"
          },
          "500": {
            "description": "erroe message"
          }
        }
      }
    }
  },
  "definitions": {
...
    "UserModel": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "info": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "DoB": {
          "type": "string",
          "format": "date-time"
        },
        "phone": {
          "type": "string"
        },
        "slack_name": {
          "type": "string"
        },
        "slack_conversational_id": {
          "type": "string"
        },
        "slack_conversational_crm_id": {
          "type": "string"
        },
        "role": {
          "type": "string",
          "enum": [
            "student",
            "user",
            "sales",
            "admin"
          ]
        },
        "repo": {
          "type": "string"
        },
        "resetPasswordExpires": {
          "type": "string",
          "format": "date-time"
        },
        "resetPasswordToken": {
          "type": "string"
        },
        "status": {
          "type": "string",
          "enum": [
            "registered",
            "active",
            "disabled"
          ]
        }
      }
    }
  }
}